refactor: split the codec and the override lookup out of state.py (#538)
state.py was 967 lines against the 800 limit, holding the JSON codec,
the atomic write path, the cross-process lock, the document mutators
and an account-scoped override lookup. It is imported by nearly the
whole codebase and by two sibling repos, so this is a pure move: every
top-level definition is byte-identical in its new home, verified by
comparing AST-extracted source segments against the base. 33 of 33.

state_codec.py takes parse_state and render_state together with their
twelve field helpers. The issue suggested moving parse alone, but every
schema field appears exactly twice - once in each direction - and the
optionality rules that keep a round trip byte-stable are only checkable
with both halves in front of you. Splitting them would have left a
contract whose two ends nobody can see at once.

conversion_overrides.py takes load_conversion_action_types. It is the
odd one out in state.py: the only thing there that resolves its own
workspace through the runtime context, the only one that joins on ad
account id, and the only one that deliberately never raises - the
opposite of the writers' strict contract.

What remains in state.py is the document layer its name promises. The
lock and the atomic write stay together; the merge semantics stay with
the mutators that implement them, because those docstrings ARE the
contract.

One behavioural difference, and only one: logger = getLogger(__name__)
is byte-identical but __name__ now resolves to state_codec, so the
tolerant read path's DEBUG records come from a different logger. Two
caplog assertions updated; nothing else in the tree filters on it.

state.py gains __all__ - not cosmetic, since mypy runs strict and
without it every existing import of a re-exported name would fail.

Four functions stay over the 50-line limit. Three are long because
their docstring is the merge contract and will stay that way.
parse_state is not permanent: its platforms loop would extract
cleanly, and was left only because a differing body would have cost
the byte-identity property that makes this change reviewable.
